requirements.txt
pycryptodome==3.21.0
scapy==2.6.0